python/requirements.txt
pymunk>=6.0
pygame>=2.0
pyzmq>=24.0
numpy>=1.21
